le-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=5.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
asyncpg>=0.29.0
aioredis>=2.0.0
aiofiles>=23.2.0
pillow>=10.0.0
pytest>=7.4.0
minio>=7.2.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
aiosqlite
python-dateutil>=2.8.0
cryptography>=41.0.0
python-magic>=0.4.27
yara-python>=4.3.1